Move integration tests to rpc/test
